agregar validaciones en controller

+++ src/controllers/products.controller.js
@@ -18,12 +18,18 @@ export const getProduct = async (req, res) => {
   try {
     const id = req.params.id;
     // console.log(id);
+    //Se valida que id no sea campos en blanco ya que de faltar el id entra en la ruta api/products/ listando todos los productos
+    if (id.trim() === "") {
+      throw new Error("No se proporciono Id");
+    }
     const product = await productsService.getProductById(id);
     // console.log(product);
     return res.status(200).json(product);
   } catch (error) {
     if (error.message === "Producto no encontrado") {
       return res.status(404).json({ error: error.message });
+    } else if (error.message === "No se proporciono Id") {
+      return res.status(400).json({ error: error.message });
     }
 
     console.error("Controller error:", error.message);
@@ -33,8 +39,15 @@ export const getProduct = async (req, res) => {
 
 export const addProduct = async (req, res) => {
   try {
+    if (!req.body) {
+      throw new Error("Falta el cuerpo en la solicitud");
+    }
     const { name, price, category } = req.body;
-    const newProduct = await productsService.addProduct({ name, price, category });
+    const newProduct = await productsService.addProduct({
+      name,
+      price,
+      category,
+    });
 
     return res
       .status(201)
@@ -42,6 +55,8 @@ export const addProduct = async (req, res) => {
   } catch (error) {
     if (error.message === "Faltan campos obligatorios") {
       return res.status(400).json({ error: error.message });
+    } else if (error.message === "Falta el cuerpo en la solicitud") {
+      return res.status(400).json(error.message);
     }
     console.error("Controller error:", error.message);
     return res.status(500).json({ error: error.message });
@@ -51,6 +66,10 @@ export const addProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const id = req.params.id;
+    if (id.trim() === "") {
+      throw new Error("No se proporciono Id");
+    }
+
     const deletedProduct = await productsService.deleteProduct(id);
 
     return res.status(200).json({
@@ -60,7 +79,10 @@ export const deleteProduct = async (req, res) => {
   } catch (error) {
     if (error.message === "Error. No se encontró el producto a eliminar") {
       return res.status(404).json(error.message);
+    } else if (error.message === "No se proporciono Id") {
+      return res.status(400).json({ error: error.message });
     }
+
     console.log("Controller error:", error.message);
     return res.status(500).json({ error: error.message });
   }
